+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Use InsecureCredentialsWrapper when authenticating with the emulator in dev 
+
 ## v0.5.1 (2023-03-30)
 
 * Allow clients to specify a custom_handler_url when creating a task, to override the default handler_url provided